Major rewrite of DB backend.
Transaction sums are now rolled up into summaries,
which are stored explicitly in the database. This
allows us to compute sums without having to read
every transaction indepedently.

This requires the ability to update several entries
in the database atomically, which has also been
added.

Some of the old functionality is still missing,
most notably reading of transactions and transaction
balances.  Coming soon.